fixed hash function, moved to Graph class from UDGraph class
